# Conflicts:
#	src/main/java/com/trs/gov/kpi/controller/IssueCountController.java
#	src/main/java/com/trs/gov/kpi/service/InfoErrorService.java
#	src/main/java/com/trs/gov/kpi/service/impl/InfoErrorServiceImpl.java
#	src/main/java/com/trs/gov/kpi/service/impl/IssueCountServiceImpl.java

+++ src/main/java/com/trs/gov/kpi/controller/FrequencyPresetController.java
@@ -1,14 +1,18 @@
 package com.trs.gov.kpi.controller;
 
+import com.trs.gov.kpi.constant.Authority;
 import com.trs.gov.kpi.constant.Constants;
 import com.trs.gov.kpi.entity.FrequencyPreset;
 import com.trs.gov.kpi.entity.exception.BizException;
+import com.trs.gov.kpi.entity.exception.RemoteException;
 import com.trs.gov.kpi.entity.requestdata.FrequencyPresetRequest;
 import com.trs.gov.kpi.entity.responsedata.ApiPageData;
 import com.trs.gov.kpi.entity.responsedata.FrequencyPresetResponse;
 import com.trs.gov.kpi.entity.responsedata.Pager;
+import com.trs.gov.kpi.ids.ContextHelper;
 import com.trs.gov.kpi.service.FrequencyPresetService;
 import com.trs.gov.kpi.service.FrequencySetupService;
+import com.trs.gov.kpi.service.outer.AuthorityService;
 import com.trs.gov.kpi.utils.PageInfoDeal;
 import com.trs.gov.kpi.utils.ParamCheckUtil;
 import lombok.extern.slf4j.Slf4j;
@@ -31,6 +35,9 @@ public class FrequencyPresetController {
     @Resource
     FrequencySetupService frequencySetupService;
 
+    @Resource
+    private AuthorityService authorityService;
+
     /**
      * 获取更新频率及预警初设的分页数据
      *
@@ -42,7 +49,11 @@ public class FrequencyPresetController {
      */
     @RequestMapping(value = "/presetfreq", method = RequestMethod.GET)
     @ResponseBody
-    public ApiPageData getPageDataBySiteId(@RequestParam("siteId") Integer siteId, Integer pageSize, Integer pageIndex) throws BizException {
+    public ApiPageData getPageDataBySiteId(@RequestParam("siteId") Integer siteId, Integer pageSize, Integer pageIndex) throws BizException, RemoteException {
+        if (!authorityService.hasRight(ContextHelper.getLoginUser().getUserName(), siteId, null, Authority.KPIWEB_INDEXSETUP_SEARCH) && !authorityService.hasRight(ContextHelper
+                .getLoginUser().getUserName(), null, null, Authority.KPIWEB_INDEXSETUP_SEARCH)) {
+            throw new BizException(Authority.NO_AUTHORITY);
+        }
         if (siteId == null) {
             log.error("Invalid parameter:  参数siteId存在null值");
             throw new BizException(Constants.INVALID_PARAMETER);
@@ -64,18 +75,22 @@ public class FrequencyPresetController {
      */
     @RequestMapping(value = "/presetfreq", method = RequestMethod.POST)
     @ResponseBody
-    public Object addFrequencyPreset(@ModelAttribute FrequencyPresetRequest request) throws BizException {
+    public Object addFrequencyPreset(@ModelAttribute FrequencyPresetRequest request) throws BizException, RemoteException {
+        if (!authorityService.hasRight(ContextHelper.getLoginUser().getUserName(), request.getSiteId(), null, Authority.KPIWEB_INDEXSETUP_ADDPREFREQ) && !authorityService.hasRight
+                (ContextHelper.getLoginUser().getUserName(), null, null, Authority.KPIWEB_INDEXSETUP_ADDPREFREQ)) {
+            throw new BizException(Authority.NO_AUTHORITY);
+        }
         if (request.getSiteId() == null || request.getUpdateFreq() == null || request.getAlertFreq() == null) {
             log.error("Invalid parameter:  参数siteId、updateFreq（更新频率）、alertFreq（预警频率）中至少一个存在null值");
             throw new BizException(Constants.INVALID_PARAMETER);
         }
-        if(request.getAlertFreq() > request.getUpdateFreq()){
+        if (request.getAlertFreq() > request.getUpdateFreq()) {
             log.error("Invalid parameter:  参数alertFreq（预警频率)大于updateFreq（更新频率）");
             throw new BizException(Constants.INVALID_PARAMETER);
         }
-        int num =  frequencyPresetService.addFrequencyPreset(request);
-        if(num == 0){
-            log.error("向站点siteId"+request.getSiteId()+"添加的预设记录已经存在");
+        int num = frequencyPresetService.addFrequencyPreset(request);
+        if (num == 0) {
+            log.error("向站点siteId" + request.getSiteId() + "添加的预设记录已经存在");
             throw new BizException("添加的预设记录已经存在");
         }
         return null;
@@ -90,19 +105,23 @@ public class FrequencyPresetController {
      */
     @RequestMapping(value = "/presetfreq", method = RequestMethod.PUT)
     @ResponseBody
-    public Object updateFrequencyPresetBySiteIdAndId(@ModelAttribute FrequencyPreset preset) throws BizException {
+    public Object updateFrequencyPresetBySiteIdAndId(@ModelAttribute FrequencyPreset preset) throws BizException, RemoteException {
+        if (!authorityService.hasRight(ContextHelper.getLoginUser().getUserName(), preset.getSiteId(), null, Authority.KPIWEB_INDEXSETUP_UPDATEPREFREQ) && !authorityService.hasRight
+                (ContextHelper.getLoginUser().getUserName(), null, null, Authority.KPIWEB_INDEXSETUP_UPDATEPREFREQ)) {
+            throw new BizException(Authority.NO_AUTHORITY);
+        }
         if (preset.getSiteId() == null || preset.getId() == null
                 || preset.getUpdateFreq() == null || preset.getAlertFreq() == null) {
             log.error("Invalid parameter:  参数siteId、id（预设记录编号）、updateFreq（更新频率）、alertFreq（预警频率）中至少一个存在null值");
             throw new BizException(Constants.INVALID_PARAMETER);
         }
-        if(preset.getAlertFreq() > preset.getUpdateFreq()){
+        if (preset.getAlertFreq() > preset.getUpdateFreq()) {
             log.error("Invalid parameter:  参数alertFreq（预警频率)大于updateFreq（更新频率）");
             throw new BizException(Constants.INVALID_PARAMETER);
         }
         int num = frequencyPresetService.updateBySiteIdAndId(preset);
-        if(num == 0){
-            log.error("在站点siteId"+preset.getSiteId()+"修改的预设记录已经存在");
+        if (num == 0) {
+            log.error("在站点siteId" + preset.getSiteId() + "修改的预设记录已经存在");
             throw new BizException("修改的预设记录已经存在");
         }
         return null;
@@ -118,14 +137,18 @@ public class FrequencyPresetController {
      */
     @RequestMapping(value = "/presetfreq", method = RequestMethod.DELETE)
     @ResponseBody
-    public Object deleteFrequencyPreset(@RequestParam("siteId") Integer siteId, @RequestParam("id") Integer id) throws BizException {
+    public Object deleteFrequencyPreset(@RequestParam("siteId") Integer siteId, @RequestParam("id") Integer id) throws BizException, RemoteException {
+        if (!authorityService.hasRight(ContextHelper.getLoginUser().getUserName(), siteId, null, Authority.KPIWEB_INDEXSETUP_DELETEPREFREQ) && !authorityService.hasRight(ContextHelper
+                .getLoginUser().getUserName(), null, null, Authority.KPIWEB_INDEXSETUP_DELETEPREFREQ)) {
+            throw new BizException(Authority.NO_AUTHORITY);
+        }
         if (siteId == null || id == null) {
             log.error("Invalid parameter:  参数siteId或者id（预设记录编号）存在null值");
             throw new BizException(Constants.INVALID_PARAMETER);
         }
-        if(!frequencySetupService.isPresetFeqUsed(siteId, id)){
+        if (!frequencySetupService.isPresetFeqUsed(siteId, id)) {
             frequencyPresetService.deleteBySiteIdAndId(siteId, id);
-        }else{
+        } else {
             log.error("Invalid parameter:  当前参数siteId和id下的记录已经被某一栏目所使用，无法删除");
             throw new BizException("当前预设记录已经被某一栏目所使用，无法删除");
         }
